Update doctest output

+++ pyvista/core/filters/data_object.py
@@ -4222,7 +4222,7 @@ class DataObjectFilters:
                          42, 43, 44, 45, 46, 47, 48, 49, 50, 51, 52, 53, 54, 55,
                          56, 57, 58, 59, 60, 61, 62, 63, 64, 65, 66, 67, 68, 69,
                          70, 71, 72, 73, 74, 75, 76, 77, 78, 79, 80, 81, 82, 83,
-                         84, 85, 86, 87, 88, 89])
+                         84, 85, 86, 87, 88, 89], dtype=int64)
         >>> surf['vtkOriginalCellIds']
         pyvista_ndarray([ 0,  0,  0,  1,  1,  1,  3,  3,  3,  2,  2,  2, 36, 36,
                          36, 37, 37, 37, 39, 39, 39, 38, 38, 38,  5,  5,  9,  9,
@@ -4230,7 +4230,7 @@ class DataObjectFilters:
                           8,  8, 12, 12, 16, 16, 20, 20, 24, 24, 28, 28, 32, 32,
                           7,  7, 11, 11, 15, 15, 19, 19, 23, 23, 27, 27, 31, 31,
                          35, 35,  6,  6, 10, 10, 14, 14, 18, 18, 22, 22, 26, 26,
-                         30, 30, 34, 34])
+                         30, 30, 34, 34], dtype=int64)
 
         Note that in the "vtkOriginalCellIds" array, the same original cells
         appears multiple times since this array represents the original cell of
